Remove winreg
Currently, we're using both the official `windows-*` with `windows-registry` crates as well as `winreg`, an older, community-maintained crate.

To unify the codebase, we follow the lead of rustup that already performed this migration (https://github.com/rust-lang/rustup/commit/bce3ed67d219a2b754857f9e231287794d8c770d). This is a prerequisite to unblock the unification of the windows-sys crate versions.

I've manually tested that `uv tool update-shell` works for adding to PATH and correctly detects when PATH was already added.