Fix non-directory in workspace on Windows
Fixes #11793

On Windows, trying to read a file inside what is not a directory but another file results in a not found error, while on unix we get a not a directory error. We check explicitly if something included in a workspace glob is a non-directory to fix the behavior on Windows.