Edition 2024 prep: Escape `r#gen` and remove redundant ref
Three edition 2021 compatible sets of changes in preparation for the edition 2025 split out from #11724.

In edition 2025, `gen` is a keyword, so we escape it as `r#gen`. `ref` and `ref mut` are not allowed anymore for `&T` and `&mut T`, so we remove them. `cargo fmt` now formats inside of macros, which the 2021 formatter doesn't undo.