fix parsing of python-platform in settings tomls
serde needs to be told where to put underscores. someone clearly noticed this
when adding attributes for schemars, but they need to be present for serde too
and then schemars gets them for free.

Strictly speaking this is a breaking change for anyone who noticed the parsing
was messed up and worked around it. We could add aliases for backcompat if we care.