Fetch keyring credentials from index URL in auth middleware
Some registries (like Azure Artifact) can require you to authenticate
separately for every package URL if you do not authenticate for the
/simple endpoint. These changes make the auth middleware aware of index
URL endpoints and attempts to fetch keyring credentials for such an
index URL when making a request to any URL it's a prefix of.

Currently, if we find a prefix index URL we attempt to fetch keyring
credentials for it. Otherwise we attempt to fetch for the full URL.

Addresses part of #4056
Closes #4583