Use index-level instead of realm-level credential caching for known indexes (#12717)
The current uv behavior is to cache credentials either at the request
URL or realm level. But in general, the expected behavior for indexes is
to apply credentials at the index level (as implemented in #12651). This
means that we also need to cache credentials at this level. Note that
when uv does not detect an index URL for a request URL, it will continue
to apply the old behavior.

Depends on #12651.