make distability opt-in
We have been claiming in our releases that we provide archives/installers for uv-build, but we only upload it as a wheel. This is because cargo-dist tries to be helpful and find all your apps, but this scales poorly to large workspaces like ours, as stuff like this slips in. So invert the default and make uv the only package dist will see until we say otherwise.