Support using `uv build-backend` in the Python backend
Provide an in-code switch to permit using the `uv build-backend` command
rather than the default `uv-build` in the Python PEP517 backend.  This
option is intended to be used by downstream packagers to provide
an option of reusing `uv` that was built already instead of having to
build a second `uv-build` executable that largely overlaps with `uv`.

Fixes #12389